Slight adjustments
