+++ src/util/motionModel.ts
@@ -31,7 +31,7 @@ import {
   normaliseLatLon,
 } from './geomath';
 import { getGnssDopKpi, Instrumentation } from './instrumentation';
-import { CameraType, ICameraFile, IMU } from 'types';
+import { CameraType, FileType, ICameraFile, IMU } from 'types';
 import { exec, ExecException, execSync } from 'child_process';
 import {
   CAMERA_TYPE,
@@ -48,6 +48,7 @@ import {
 } from 'config';
 import { DEFAULT_TIME } from './lock';
 import {
+  copyFileToUSB,
   getDateFromFilename,
   getDateFromUnicodeTimestamp,
   promiseWithTimeout,
@@ -71,11 +72,6 @@ export const MIN_PER_FRAME_BYTES = 25 * 1000;
 
 const MIN_DISTANCE_BETWEEN_FRAMES = 1;
 const MIN_TIME_BETWEEN_FRAMES = 33; // Max 30fps
-
-const enum  FILETYPE {
-  GNSS = 'GNSS',
-  IMU = 'IMU',
-}
 
 const defaultImu = {
   threshold: 0.05,
@@ -307,54 +303,6 @@ let emptyIterationCounter = 0;
 let prevGnssFile = '';
 let prevGpsRecord: GNSS | undefined = undefined;
 
-const createFileNameForFAT32 = (gnssFile: string, fileCreationDate: string) => {
-  //We exclude all dots and colons from the filename and replace with - for FAT32 compatibility
-  const parts = gnssFile.split('.');
-  if (parts.length > 1) {
-    const lastPart = parts.pop(); // Remove the last part
-    const replacedString = parts.join('-') + '.' + lastPart;
-    return replacedString;
-  }
-  return '';
-};
-
-const copyFileToUSB = async (fileName: string, fileType: FILETYPE) => {
-
-  const execAsync = promisify(exec);
-
-  // Replace all colons and periods with dashes to make them compatible with FAT32
-  const fileNameForFAT32 = fileName.split('/').pop()?.replace(/:/g, '-');
-
-  if (fileNameForFAT32) {
-    // Get filename of GNSS file
-
-    const usbConnected = existsSync(USB_WRITE_PATH);
-
-    if (usbConnected) {
-
-      const fileCreationDate = fileNameForFAT32.split('T')[0];
-      const destinationFileName = createFileNameForFAT32(fileNameForFAT32, fileCreationDate);
-
-      if (destinationFileName) {
-        const destinationFilePath = USB_WRITE_PATH + '/' + fileCreationDate + '/'+ fileType+ '/' + destinationFileName;
-
-        try {
-          await fs.mkdirSync(path.join(USB_WRITE_PATH, fileCreationDate, fileType));
-        }
-        catch (err) {
-          if (!((err as NodeJS.ErrnoException).code === 'EEXIST')) {
-            console.error(`Error creating directory for ${fileType} file storage: ${err}`);
-          }
-        }
-        const result = await execAsync(`cp ${fileName} ${destinationFilePath}`);
-        if (result.stderr) {
-          console.error(`Error copying ${fileType} file to USB Stick: ${result.stderr}`);
-        }
-      }
-    }
-  }
-};
-
 export const getNextGnss = (): Promise<GnssMetadata[][]> => {
   return new Promise(async (resolve, reject) => {
 
@@ -364,7 +312,7 @@ export const getNextGnss = (): Promise<GnssMetadata[][]> => {
       console.log('Last file is ' + prevGnssFile);
       pathToGpsFile = await getNextGnssName();
       if(pathToGpsFile !== prevGnssFile){
-        copyFileToUSB(pathToGpsFile, FILETYPE.GNSS);
+        copyFileToUSB(pathToGpsFile, FileType.GNSS);
       }
 
       console.log('Next file is: ' + pathToGpsFile);
@@ -698,7 +646,7 @@ export const getNextImu = (gnss: GnssMetadata[]): Promise<ImuMetadata> => {
                   encoding: 'utf-8',
                 });
 
-                copyFileToUSB(IMU_ROOT_FOLDER + '/' + imuFile, FILETYPE.IMU);
+                copyFileToUSB(IMU_ROOT_FOLDER + '/' + imuFile, FileType.IMU);
 
                 let output = '';
                 try {
